build: drop AM_C_PROTOTYPES.
This macro is deprecated as of Automake 1.12 and can no longer be
used. Most likely is not what was intended to be used.

Make sure to explicitly check for string.h after this change as it was
implied by the macro before.

See http://www.flameeyes.eu/autotools-mythbuster/forwardporting/automake.html
for details.