build: look for pthread_create when finding pthread library.
The pthread_mutex_lock function is defined as a stub in libc (for
glibc), so it'll not find the need for -lpthread.

The only reason why flowgrind would link correctly otherwise is that
other libraries bring pthread in, but this check needs to be fixed
when enabling neither pcap nor gsl.