Unique identifier for intermediated interval report columns
In the current implementation it was very error-prone to add a new column to
the intermediated interval report, because of the state of intermediated
interval report was widely spread around different arrays with different sizes.

Now we have 'one' array that tracks everything: the name and unit of the column
as well as the status (visibility, column width). Moreover, each column has now
a unique identifier. We can now easily address ever column directly instead of
an error-prone looping.