flowgrind_stop - clean up usage and option parsing
This patch cleans up the option parsing and usage. The usage and the error
messages displayed during the command line parsing are inspired by the
coreutils

Flowgrind's argument parsing is still far away from being clean. Most of the
problems arise from shortcomings of getopt_long(). In a further refinement of
the option parser we should think of using another library. At a first glance
GNU's 'arg_parser' seems promising. Other possibilities are GNU's 'argp',
and the parser from git (which is used by perf for example), which are both
may not lightweight enough