make RPC api independent of libpcap availability
Before this patch, controller and daemon had to be compiled
with the same featureset reagarding libpcap, that is either
both had to compiled with libcpap or none.

This patch fixes that. The RPCs itself will now always contain
the corresponding options, and the daemon will always be able
to receive them.

However, if the controller asks for dumps, but the daemon
is not compiled with libpcap, it will complain.

As a side effect, we get rid of a lot of #ifdefs