flowgrindd: limit flows a daemon instance is willing to handle
Currently we use portable select() API, which is limited to the number
of bits in a fd_set. Thus we cannot allow FD numbers greater than 1024
in one daemon instance.
Because of reasons we may use two FDs for one destination handled by the
daemon. Thus this patch limits the number of flows a daemon instance will
handle to half of FD_SETSIZE.
FD_SETSIZE is 1024 on linux.

Closes #108